+++ app/src/main/java/com/gooduckrefactoring/data/home/HomeRepository.kt
@@ -1,29 +1,51 @@
-package com.gooduckrefactoring.repository.remote
+package com.gooduckrefactoring.data.home
 
 import android.util.Log
 import com.gooduckrefactoring.api.RetrofitInstance
+import com.gooduckrefactoring.data.Result
 import com.gooduckrefactoring.dto.BasicResponse
-import com.gooduckrefactoring.util.Result
 import org.json.JSONObject
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class CategoryRepository() {
+class HomeRepository() {
 
     companion object {
-        private var instance: CategoryRepository? = null
+        private var instance: HomeRepository? = null
 
-        fun getInstance(): CategoryRepository? { // singleton pattern
+        fun getInstance(): HomeRepository? { // singleton pattern
             if (instance == null) {
-                instance = CategoryRepository()
+                instance = HomeRepository()
             }
             return instance
         }
     }
 
-    suspend fun getRequestAllCategory(result: (Result<BasicResponse>) -> Unit) {
-        RetrofitInstance.apiList.getRequestAllCategory().enqueue(object : Callback<BasicResponse> {
+    suspend fun getRequestBanner(result: (Result<BasicResponse>) -> Unit) {
+        RetrofitInstance.apiList.getRequestBanner().enqueue(object : Callback<BasicResponse> {
+            override fun onResponse(call: Call<BasicResponse>, response: Response<BasicResponse>) {
+                if (response.isSuccessful) {
+                    result(Result.Success(response.body()!!))
+//                    Log.d("login_code",  response.body()!!.toString())
+                } else {
+                    val errorBodyStr = response.errorBody()!!.string()
+                    val jsonObj = JSONObject(errorBodyStr)
+                    val message = jsonObj.getString("message")
+                    val code = jsonObj.getString("code")
+//                    result(Result.Error(message))
+                    Log.d("login_code", code + message)
+                }
+            }
+            override fun onFailure(call: Call<BasicResponse>, t: Throwable) {
+                t.message?.let { Log.d("########", it) }
+//                result(Result.Error(it))
+            }
+        })
+
+    }
+    suspend fun getRequestAllProduct(result: (Result<BasicResponse>) -> Unit) {
+        RetrofitInstance.apiList.getRequestAllProduct().enqueue(object : Callback<BasicResponse> {
             override fun onResponse(call: Call<BasicResponse>, response: Response<BasicResponse>) {
                 if (response.isSuccessful) {
                     result(Result.Success(response.body()!!))
