+++ app/src/main/java/com/gooduckrefactoring/data/user/UserRepository.kt
@@ -1,37 +1,29 @@
-package com.gooduckrefactoring.repository.remote
+package com.gooduckrefactoring.data.user
 
 import android.util.Log
 import com.gooduckrefactoring.api.RetrofitInstance
+import com.gooduckrefactoring.data.Result
 import com.gooduckrefactoring.dto.BasicResponse
-import com.gooduckrefactoring.util.Result
 import com.nepplus.gooduck.models.UserData
 import org.json.JSONObject
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class UserRepository() {
+class UserRepository(private val userDatasource : UserDatasource) {
 
     companion object {
         private var instance: UserRepository? = null
 
         fun getInstance(): UserRepository? { // singleton pattern
-            if (instance == null) instance = UserRepository()
+            val userDatasource = UserDatasource()
+            if (instance == null) instance = UserRepository(userDatasource)
             return instance
         }
     }
 
     suspend fun getRequestMyInfo(result: (Result<UserData>) -> Unit) {
-        RetrofitInstance.apiList.getRequestMyInfo().enqueue(object : Callback<BasicResponse> {
-            override fun onResponse(call: Call<BasicResponse>, response: Response<BasicResponse>) {
-                if (response.isSuccessful) {
-                    result(Result.Success(response.body()!!.data!!.user))
-                }
-            }
-            override fun onFailure(call: Call<BasicResponse>, t: Throwable) {
-                t.message?.let { Log.d("########", it) }
-            }
-        })
+        userDatasource.getRequestMyInfo(result)
     }
 
     suspend fun postRequestLogin(email: String, pw: String, result: (Result<BasicResponse>) -> Unit) {
