+++ app/src/main/java/com/gooduckrefactoring/util/ContextUtil.kt
@@ -35,39 +35,6 @@ class ContextUtil {
         }
 
 
-        fun setRecentSearch(context: Context, values: ArrayList<String>) {
-            val pref = context.getSharedPreferences(prefName, Context.MODE_PRIVATE)
-            val editor = pref.edit()
-            val a = JSONArray()
-            for (i in 0 until values.size) {
-                a.put(values[i])
-            }
-            if (!values.isEmpty()) {
-                editor.putString(RECENT_SEARCH, a.toString())
-            } else {
-                editor.putString(RECENT_SEARCH, null)
-            }
-            editor.apply()
-        }
-
-        fun getRecentSearch(context: Context): ArrayList<String>? {
-            val pref = context.getSharedPreferences(prefName, Context.MODE_PRIVATE)
-            val json = pref.getString(RECENT_SEARCH, null)
-            val list = ArrayList<String>()
-            if (json != null) {
-                try {
-                    val a = JSONArray(json)
-                    for (i in 0 until a.length()) {
-                        val url = a.optString(i)
-                        list.add(url)
-                    }
-                } catch (e: JSONException) {
-                    e.printStackTrace()
-                }
-            }
-            return list
-        }
-
 
         fun clear(context: Context){
             val pref = context.getSharedPreferences(prefName, Context.MODE_PRIVATE)
