+++ app/src/main/java/com/gooduckrefactoring/viewmodel/SearchViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.gooduckrefactoring.repository.remote.CategoryRepository
-import com.gooduckrefactoring.util.Result
+import com.gooduckrefactoring.data.category.CategoryRepository
+import com.gooduckrefactoring.data.Result
 import com.nepplus.gooduck.models.Category
 import kotlinx.coroutines.launch
