+++ app/src/main/java/com/gooduckrefactoring/viewmodel/SmsViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.gooduckrefactoring.dto.BasicResponse
-import com.gooduckrefactoring.repository.remote.UserRepository
+import com.gooduckrefactoring.data.user.UserRepository
 import java.util.regex.Pattern
